Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/child-src/frame-blocked.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Security-Policy" content="child-src 'none'">
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
</head>
<body>
<p>This tests that an &lt;iframe&gt; load is blocked when using Content Security Policy <code>child-src 'none'</code>. This test PASSED if there is no JavaScript alert.</p>
<iframe src="http://127.0.0.1:8000/security/contentSecurityPolicy/resources/alert-fail.html"></iframe>
</body>
</html>
